review: EmojiReactionTopN 컴포넌트 다른 jsx로 분리
